move record-mode module
