+++ src/js/engine/record-mode.js
